feat: normalize types, fix bugs, and refactor the code.

+++ src/StardustSandbox/Colors/TemperatureColorRange.cs
@@ -2,10 +2,10 @@
 
 namespace StardustSandbox.Colors
 {
-    internal readonly struct TemperatureColorRange(double minimumTemperature, double maximumTemperature, Color color)
+    internal readonly struct TemperatureColorRange(float minimumTemperature, float maximumTemperature, Color color)
     {
-        internal readonly double MinimumTemperature => minimumTemperature;
-        internal readonly double MaximumTemperature => maximumTemperature;
+        internal readonly float MinimumTemperature => minimumTemperature;
+        internal readonly float MaximumTemperature => maximumTemperature;
         internal readonly Color Color => color;
     }
 }
