+++ src/StardustSandbox/Constants/SimulationConstants.cs
@@ -2,8 +2,8 @@
 {
     internal static class SimulationConstants
     {
-        internal const double NORMAL_SPEED_DELAY_SECONDS = 0.1;
-        internal const double FAST_SPEED_DELAY_SECONDS = 0.05;
-        internal const double VERY_FAST_SPEED_DELAY_SECONDS = 0.01;
+        internal const float NORMAL_SPEED_DELAY_SECONDS = 0.1f;
+        internal const float FAST_SPEED_DELAY_SECONDS = 0.05f;
+        internal const float VERY_FAST_SPEED_DELAY_SECONDS = 0.01f;
     }
 }
