+++ src/StardustSandbox/Elements/Solids/Immovables/Iron.cs
@@ -4,9 +4,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class Iron : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue > 1200)
+            if (currentValue > 1200.0f)
             {
                 context.ReplaceElement(ElementIndex.Lava);
                 context.SetStoredElement(ElementIndex.Iron);
