+++ src/StardustSandbox/Elements/Solids/Immovables/RedBrick.cs
@@ -4,9 +4,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class RedBrick : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue >= 1727)
+            if (currentValue >= 1727.0f)
             {
                 context.ReplaceElement(ElementIndex.Lava);
                 context.SetStoredElement(ElementIndex.RedBrick);
