+++ src/StardustSandbox/Elements/Solids/Immovables/Wood.cs
@@ -5,9 +5,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class Wood : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue >= 300)
+            if (currentValue >= 300.0f)
             {
                 if (SSRandom.Chance(65))
                 {
